Added (namespace) version to Component annotation
Added more checks to make sure the namespace is set
to 1.1.0 when a 1.1.0 feature is used. Especially when
package private or private methods are used.

Changed from the default namespace for the component
element to prefix only, leaving the children elements
unqualified. This is how it should be according to the
spec :-(

Added test cases for proper handling of the new
namespace handling.