fix findbugs warning: Check for sign of bitwise operation
Check for sign of bitwise operation in
  bndtools.editor.BndEditor.resourceChanged(IResourceChangeEvent)

This method compares an expression such as
  ((event.detail & SWT.SELECTED) > 0)
Using bit arithmetic and then comparing with the greater than
operator can lead to unexpected results (of course depending
on the value of SWT.SELECTED). If SWT.SELECTED is a negative
number, this is a candidate for a bug. Even when SWT.SELECTED
is not negative, it seems good practice to use '!= 0' instead
of '> 0'.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>