fix findbugs warning: Comparison of String parameter using == or !=
Comparison of String parameter using == or != in
  bndtools.editor.completion.BndSourceViewerConfiguration.getDefaultPrefixes(ISourceViewer, String)

This code compares a java.lang.String parameter for reference
equality using the == or != operators. Requiring callers to
pass only String constants or interned strings to a method is
unnecessarily fragile, and rarely leads to measurable performance
gains. Consider using the equals(Object) method instead.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>