fix findbugs warning: Possible null pointer dereference
Possible null pointer dereference of editResult in
  bndtools.editor.common.MapEntryCellModifier.modify(Object, String, Object)

There is a branch of statement that, if executed, guarantees that a
null value will be dereferenced, which would generate a
NullPointerException when the code is executed. Of course, the
problem might be that the branch or statement is infeasible and that
the null pointer exception can't ever be executed; deciding that is
beyond the ability of FindBugs.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>