fix findbugs warning: is or uses a map or set of URLs
bndtools.editor.project.OSGiFrameworkLabelProvider.images is or
uses a map or set of URLs, which can be a performance hog

This method or field is or uses a Map or Set of URLs. Since both
the equals and hashCode method of URL perform domain name resolution,
this can result in a big performance hit.
See http://michaelscharf.blogspot.com/2006/11/javaneturlequals-and-hashcode-make.html
for more information. Consider using java.net.URI instead.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>