fix FB warning: Dead store
Dead store to cmd in aQute.bnd.maven.MavenDeployCmd.run(String[], int)

This instruction assigns a value to a local variable, but the value is
not read or used in any subsequent instruction. Often, this indicates
an error, because the value computed is never used.

Note that Sun's javac compiler often generates dead stores for final
local variables. Because FindBugs is a bytecode-based tool, there is
no easy way to eliminate these false positives.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>