fix FB warning: inefficient use of keySet iterator
aQute.lib.getopt.CommandLine.getOptions(Class, List) makes
inefficient use of keySet iterator instead of entrySet iterator

This method accesses the value of a Map entry, using a key that
was retrieved from a keySet iterator. It is more efficient to use
an iterator on the entrySet of the map, to avoid the Map.get(key)
lookup.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>