fix FB warning: Format-string
Format-string method String.format(String, Object[]) called with
format string "Failed to reset Workspace OBR provider: could not
delete index file " wants 0 arguments but is given 1 in
bndtools.WorkspaceObrProvider.reset()

A format-string method with a variable number of arguments is
called, but more arguments are passed than are actually used by
the format string. This won't cause a runtime exception, but the
code may be silently omitting information that was intended to be
included in the formatted string.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>