I implemented support for BND to handle bundle projects at various levels in a workspace.
To use, add the "-projectsearch" BND directive to cnf/build.bnd or cnf/ext/*.bnd

An example:
   -projectsearch: .;depth=1,dir;depth=1
This tells BND to search for projects under the workspace and any projects under /dir. But only search one level down into each. The depth parameter defaults to 0, which means that it doesn't recurse down into another level of directories.
If -projectsearch is not specified, the default is the old behavior, which should be equivalent to ".;depth=1"