#196 Changed resource that was a CPE failed the suffix method
The cause was that there was an invalid assumption that ClassPath Entries (CPEs) were directories. So there was a failure when a CPE had changed and was part of the resource change set. 

The getClassPathEntrySuffix therefore now returns "" signalling a CPE change. Maybe we need to check the subset of packages in that CPE and only return true when we actually include that but that seems a bit over the top for now.