Changed the release target.
The release target was doing a recursive build which
meant n(n-1) (ok, a bit less) projects were released
and indexed. It now ensures all projects are
build then does a local release for each project. After
all locals are released, this project is released
This target is mainly, if not only, intended for the
diet project and is called as such from the top build.