Enable source attachments on libraries in the 'Bnd Bundle Classpath'
The current implementation stores source attachment data (if there is any) for each library exposed via the BndClasspathContainer. The attachment paths are persisted to a properties file per project.

There is also partial support for eclipse style source bundles, that is bundles that are of the form <bsn>.source are attached to bundles as source if there is not already a source attachment set.