Added a macro that gets a repository digest
A repository digest is a unique digest representing
the contents of a repo. It is calculated by taking the
sha-1 of its contents, sorting them, removing duplicates
and then digesting them with sha-1. The macro
visits all repositories and any repo with a getDigest
method is added to the list.

jim can use this to calculate the dependencies.