Fix a findbugs warning
Printer.java:87
RCN_REDUNDANT_NULLCHECK_WOULD_HAVE_BEEN_A_NPE, Priority: High

Nullcheck of m at line 87 of value previously dereferenced in
bndtools.jareditor.internal.Printer.doPrint(File, int)

A value is checked here to see whether it is null, but this value can't
be null because it was previously dereferenced and if it were null a null
pointer exception would have occurred at the earlier dereference.
Essentially, this code and the previous dereference disagree as to whether
this value is allowed to be null. Either the check is redundant or the
previous dereference is erroneous.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>