[lib] Added a PersistentMap
This map is a <very> low performance Map that uses the file system as storage. The keys of this map are the names of files in a directory that is given to the map. The values must be usable in the JSONCodec, they are written and read as used. They values are held with soft references, allowing them to be gc'ed