[test] The project test clean up was wrong
The original tests polluted the directories in the workspace. Ferry added a method to clean up these artifacts but this created a very tight coupling between today's layout and future changes.

Changed all the tests back to be non static (why static???) and added a method that copies a workspace to a tmp dir. This tmp dir is then deleted at startup/teardown