[core] Added profile support for 1.8 (experimental) You can enable this by -eeprofile: auto
If you want to specify your own profiles (works for all EEs):

	-eeprofile: profile1="a,b,c", profile2="d,e,f"

Where a..f are java.* packages. Profiles must not overlap. If one or more profiles
are found, the highest lexically is asserted on the filter.

It is experimental because I think the profiles are the wrong dimension, we can just better use unique names.