#868 The dependency building was broken
I think a patch to improve the performance of building broke the dependencies. If a A depends on B (version=project or version=latest) then a change in class files did not cause a rebuild of any projects that depended on this project.

This change will force a rebuild if any class file in the output directory of a dependent project is changed.

PLEASE VERIFY

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>