[file & macros] Erroneously adapted to Windows
Ferry had made an attempt to fix the lsa/lsr test cases to make them run on windows. However, this did not work (the replace failed with a backslash as separator) and you cannot use \ in macros this way, that would require double escaping of the backslashes. So this test was not tested. Anyway, bnd always uses forward slashes. However, the result was returned with backslashes.