#841 Fix headless build plugins on Windows
BundleResourceCopier doesn't take into account that the Windows path separator
is a backslash.  Bundle.getEntryPaths and Bundle.getEntry only deal with
forward slashes, meaning that the GradleHeadlessBuildPlugin and
AntHeadlessBuildPlugin both fail in their respective setup() methods.

So for paths that will end up being passed to getEntryPaths or getEntry, we
simply (and perhaps naively) replace backslashes with forward slashes.