JPM View: Support the platform default browser
SWT on Windows uses Internet Explorer out of the box, and while I
believe it is possible to get WebKit or XULRunner working with some
downloads/configuration changes, it is nice to have this working by
default instead of silently failing.

Add a "Platform Default" option to the JPM browser selection and make it
the default.  This should work on most default installs across platforms
and if it doesn't the user can choose a specific browser engine if they
choose to do so.

Signed-off-by: Sean Bright <sean@malleable.com>