[project] BSN naming calculation caused unexpected results
The calculation of the default BSN of a sub bundle looked at the directory name and when the sub-bundle started with that name, it did not prefix the directory name. I.e. if you have a bundle prefix.bnd in a project prefix, the BSN would be prefix, not prefix.prefix. However, this caused lots of confusion (see bndtools bug nr 903). 

So we now do not have this feature anymore.

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>