jareditor: Add bndlib as a plugin for the jareditor feature
It turns out the jareditor plugin has long had a dependency on bndlib.
Because bndlib packages can add dependencies to other packages, the
Conditional-Package list in jareditor will fail to include them and
import package references will be generated. This happened some time ago
for the aQute.service.reporter package and just recently for the
org.osgi.service.metatype annotation package. The solution is for
jareditor to simply import any necessary packages from bndlib (as it
currently does) and for bndlib to be part of the jareditor feature.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>