[baseline] #652 Changes when nested annotations are used
Nested annotations (in array and single) form were not treated as nested annotations but were handled with toString. However, if a nested annotation contained an array, then the array was not properly printed.

So there is now a single method that takes an Annotation and turns it into an ANNOTATED element. The children will either be a PROPERTY or ANNOTATED. If a child is an array, then each element is treated this way but the key is suffixed with its index.

Simple properties are now also enclosed in single quote when string. The current format could not see the difference if an annotation had changed from int to string. Assuming that one number to another should not matter.

For example:

    ANNOTATED test.annotations.diff.Outer (CHANGED/CHANGED)
     ANNOTATED test.annotations.diff.Inner (CHANGED/CHANGED)
      PROPERTY value.0='1' (CHANGED/CHANGED)
      PROPERTY value.1='2' (CHANGED/CHANGED)
     PROPERTY x.0=1 (CHANGED/CHANGED)
     PROPERTY x.1=2 (CHANGED/CHANGED)
     PROPERTY x.2=3 (CHANGED/CHANGED)

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>