Fix transitive jar file references issues
Handles case 1 where:
Java Project A outputs a.jar containing package A
OSGI Project B has a version=file reference to a.jar and includes package A in export-packages.
OSGI Project C has buildpath reference to B and uses something in package A.

Without this change, Project C couldn't build due to
"not accessible due to restriction on required project" errors

Handles case described in bug #937

Basically, to maintain the no-transitive dependencies philosophy of bnd,
we shouldn't use transitive dependencies for the eclipse classpath. This
commit adjusts where transitive dependencies are used.
However there are 2 cases where we must
1) when a version=file reference has a package that will be exported by
the current project. This package becomes an available transitive
dependency for a user of the current project. All other packages are
made available to the current project, but are not transitive.
2) when a standard buildpath reference has a package that will be exported
by the current project. This package becomes an available transitive
dependency for a user of the current project. Not all other packages are
made available to the current project, only exported packages from the
referenced project, but they are not transitive dependencies.

The part of this that wasn't quite right before was that all project
references were turning into transitive dependencies, which caused
issues.. Especially with subbundles.

Signed-off-by: Carter Smithhart <carter.smithhart@gmail.com>