Adjust newaccessrules classpath entries for Container.TYPE.PROJECT
Make a library and a project entry for each buildpath entry that resolves
to an eclipse project.

This seems to address the source lookup and source refactoring
problems discussed in #1005. And it seems to have all of the benefits
of what the new access rules are bringing.

More bake time now...

Signed-off-by: Carter Smithhart <carter.smithhart@gmail.com>