builder: Set project build order using project dynamic references
Setting the workspace build order has a problem if the user has the
workspace preference for default build order set. Then the code would
not be able to see the default build order so that it could modify the
bnd projects and preserve the non-bnd projects.

This change removes setting the workspace build order with setting
project dynamic references based upon the calculated project depends on.
The allows eclipse to use the default workspace build order based upon
the additional project dynamic references calculated by bnd and set by
bndtools.

We need to use the complete project depends (no removal of the
dependencies of dependencies) so that Eclipse will rebuild the project
if any of its dependencies change. In removing the removal of
dependencies of dependencies, we remove a bug where the original
dependsOn collection from bnd Project was mutated.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>