[builder] Changed strategy for source code
The previous model added the project before the JAR with forbidden access and the project as source. The JAR then had full access while its source path was also the project. This is now changed. The referred project now has the identical rules as the JARs. The project now has the project as source code and the JAR now has the JAR as source code.

This means that if a class is in the project, it will be properly picked up. However, if you copy code from other projects in the generated JAR and include source code, this will also be found.

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>