Special treatment for artifacts with version numbers consisting only of release numbers
Google Guava for example has a released artifact with version string r03, which will cause and IllegalArgumentException in Version.parseVersion() when called from MvnVersion.parseString().
This fix interprets version strings of the form r<digits> as a version with major and minor parts set to 0 and the release number as the micro version.

Signed-off-by: Adrian Rumpold <a.rumpold@gmail.com>