classpath: Remove setting of classpath error markers
The container initializer will now put all errors in the model and the
incremental builder will be responsible for setting the error markers.

Currently the incremental builder loses the errors, so it needs to be
changed. @pkriens is working on this change. The incremental builder
also needs to make sure Project.prepare() is run each incremental build
to generate any classpath errors that exist into the model. Otherwise
the classpath errors will be lost if the model is cleared each build.

Since the incremental builder is now responsible for all error markers,
the bnd classpath problem marker type is also deleted.

Fixes https://github.com/bndtools/bndtools/issues/1037

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>