[builder] Diverse optimizations and ensuring correct error reporting
The builder now separates the change of the workspace or project setup from the changes in its resources. Only if the setup is changed do we calculate the dependencies and paths.

Two additional error markers have been added. Any problem that is reported when changing the setup is using a path problem marker. This marker is only deleted in the setup change block and independent of the build markers.

Additionally, any errors in the workspace setup are now reported on the cnf project, the build.bnd file.

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>