cnfwatcher: Fix NPE when cnf project was deleted from the workspace
The code did not check that the IProject for cnf was valid before
attempting to mark it. The fix avoids using hard coded name for cnf and
uses the findCnfProject method to locate the valid IProject for cnf, if
there is one.

The fix also ensures only a single instance of CnfWatcher can be created.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>