#761 ${tstamp} inconsistent in one build
Since the macro always reads from the current time an unlucky build can fail because the time between the setting and verification in bnd can cross a minute.

This change makes the tstamp macro verify a @tstamp long property set to the time to be used for tstamps. If it is not set, the current time is used.

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>