gradle plugin: Add gradle support for non-Workspace environments
A new Bundle task type is defined which extends the Jar task type and
used the bnd Builder class to make the jar into a bundle. An
biz.aQute.bnd.builder plugin is defined which, when applied to a project
modifies the jar task to use the bnd Builder class to make the jar into
a bundle.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>