classpath container: suggest container update needs to use a timestamp
The old impl uses the JarInfo cache, but updating one container will
update the cache so that a later container would not see that it needed
to be updated. So now we hold a lastModified timestamp in the container
that holds the timestamp of the newest item in the container. The
suggest change method will compare the timestamp of the container to the
timestamps of the container items and suggest an update if any item is
newer than the container timestamp.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>