[builder] #840 Negation in Private package causes unused warnings
When using negation with Private-Package a warning was issued that the negation was never used. Export-Package was ok.

cause: The negation bailed out of the loop before the ‘used’ flag was set, this caused the filter to end up in the unused set.

Why did this not happen in Export-Package? Good question. ExportPackage had the same problem but unused packages are not reported in the builder since they are also detected during the final verification, unlike Private Packages.

#840 Fixed

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>