[resolve] Use the containers instead of assuming that version in a path is a valid version
The analysis of the -run/testpath for capabilities assumed that the version was a valid version and the bsn was a valid bsn. However, the path can actually contain some variations that are interpreted by the project (latest, project, snapshot). So the code now gets the containers (which will now also take care of libraries) and then uses the manifest of these entries.

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>