bnd-maven-plugin: Handle projects with no source code
The project might just want to wrap some other jars. This fix will test
for the target/classes dir and if it is not a directory, it will not
add it to bnd's classpath or the -includeresource directive.

Also added a new test project to the integration-test to test building
a project with no source code.

Fixes https://github.com/bndtools/bnd/issues/925

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>