gradle plugin: jar task rebuild for -dependson project's artifact changes
The jar task would rebuild if any of the -buildpath dependencies change
but when a -dependson dependency changes, the jar task would not rebuild.
The OSGi build uses -dependson for packaging projects like osgi.ct to
depend up on the various ct projects. But when a ct project rebuilt, the
osgi.ct project would not rebuild the packaging jar. This change means
that the jar task will rebuild the jar of any dependency in -buildpath
or -dependson changes.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>