[resolve] Restructured and cleaned up Resolver
The resolver had gotten quite messy. When adding the distro’s it turned out that it was hard to see if a GenericResolveContext was in charge or the BndResolveContext. Many fields were shared and set by methods with sideEffects

Added an AbstractResolveContext that is pure, it does not share fields. Most important, the model has been reduced from many half defined variables to 2 resources:

inputResource -> contains the initial requirements
system        -> all system capabilities (framework, runpath, distro, etc)

The BndrunResolveContext now programs the base class in its init method. All its methods calculate requirements and capabilities that are added to their proper place

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>