[cleanup] Removed indexer because it overlaps with RepoIndex
This was a futile attempt to replicate RepoIndex and should not be used. We should however, try this so that bnd can replace RepoIndex. The current RepoIndex stupidly tries to abstract itself from bnd, replicating many classes. 

The aQute.bnd.resource package now has significant support for parsing files and generating index files. (see addManifest on resource). These classes do not yet support the non-manifest parsing but one should question how useful that is. The manifest should become the sole definition in the end.

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>