[resource] Versions
In a capability a version MUST be a org.osgi.framework.Version object (or a collection thereof). In Attrs, a version MUST be a aQute.bnd.version.Version :-)

This is now checked and enforced. This caused some exceptions to be thrown which rippled through to quite some code. Unfortunately, some code idly assumes no exceptions are thrown :-(



Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>