[resource] Additional utilities for the resource package
- ResourceImpl is now comparable
- corrected a missing ) in filter
- Added support to add an EE to a resource
- Added Resolution enum

Had to move to 2.0.0 but this is bnd 3.0.0 & I do not think there are any external users yet. Might change though

If I had time I would add the other enums & support. We could then soon throw away repoindex because it would be much nicer to do this in bnd.





Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>