[build] Multiple sub bundles did not show up in Workspace Repository
It turned out that each sub bundle build was clearing the map with the other sub bundles. This caused only the last sub bundle to remain in that map. I.e. the workspace repository only showed the last sub bundle.

The project now clears the map before the overall build. Sub bundle builds remove and add their name + version to  this map.

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>