gradle: Use an actual bnd file for the non-workspace plugin
Processor really wants a properties file. So we now
write the manifest properties and the specified bnd file into a temp
bnd file which we use to set the Builder's properties file.

We also clean up the sub builders check since we don't use the specified
bnd file as the Builder's properties file.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>