default content: Use setJar to set default content
This is the same way the gradle and maven plugins do it. So we are
consistent in how the default content is added to the builder.

doDefaults is backed out since we need the output folder created
which is done by prepare() which called doDefaults causing a cycle. So
we instead use the existing startBuild method which delays the setJar
call until immediately before the build which is better since
getSubBundles is called many times when we are not actually building.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>