[build] Expand the bundle class path
This patch is a replacement for #1177. It expands the BCP when the expand-bcp attribute is set on a repo entry. 

If this flag is set, the Bundle-Classpath entries are stored in a directory that is in the same directory as the file. It starts with ‘.’ to prevent problems with some repositories. The file names are numbered to indicate their position on the BCP and make them unique. (We really want a file name here since it is displayed in bndtools.)

Access is cached so the cost should be doable.

I decided to redo the PR because:

1) Wanted to store the temp files very close to the original files so they get deleted
2) The PR looked very costly in CPU; it required always the calculations of the SHA.



Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>