Maven Bnd Repository
This commit introduces the Maven Bnd Repository. This repository was
the first project that heavily used the HttpClient so a number of changes
are in the Http Client code and its API package. Mainly TaggedData has
grown. Further, methods were added to aQute.lib for working with
hex, strings, and urls.

It slightly modifies the Project class. Project now passes a context in
put so that the Maven bnd Repo can pickup the -maven-release function.
It also added a -installrepo instruction. Any build will be released as
to the repos listed in there.

It also accidentally includes the fix to #1352

Fixes #1352

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>
Signed-off-by: BJ Hargrave <bj@bjhargrave.com>