[macro] Allow filter expression in if
The ${if} macro assumes the first argument is a filter expression if it starts with a ( and and ends with ). It will then compile this expression and use the current domain as context for the filter.

Since the filter tends to work on real objects and we have only strings, you can suffix a key with [] to treat it as an array by doing the standard comma separator thing.

The filter was extended to support !=, >, and <

So now you can do things like:

   ${if;(&(Bundle-Version=*)(Bundle-SymbolicName=*)(foo>3));YES;NO}



Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>