[maven] Sources don’t work
I see there is a date/time problem with the file dates. The JAR write sets the file date to bnd highest last modified resource in the manifest and that somehow disagrees with file date of the binary jar. When we build the class path and ask for the binary file we see the sources are older and ignore them. :-(

The file date of the sources file is now > binary JAR.



Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>