gradle: Fix bnd extension property lookup for Gradle 2.14
After some more testing, I need to change the strategy to properly
handle contained extensions that someone might set in bnd. So, instead
of returning the ext property or the extension object, we now throw a
MissingPropertyException if the name is 'ext' or mapped to an extra
property or a contained extension. Then Gradle will complete the lookup
of the value.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>