container: Cannot defer initialization of bnd classpath container
The change to defer the initialization of the bnd class path container
to after the bnd workspace is initialized can result in the failure to
establish the bnd class path container for a java project. When
initialize is called by Eclipse, the code must call setClasspathContainer
otherwise Eclipse will decide the java project does not have a bnd
class path container. So we always start by calling setClasspathContainer
to set an empty class path and establish the bnd class path container.
Then, after the bnd workspace is initialized, we are called again and
will set the actual class path for the container.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>