classpath container: Serialize to allow restore before workspace init
We use a classpath container serializer/deserializer helper from M2E to
serialize the BndContainer when it is set. During initialization of the
classpath container, if the workspace is initialized, we do the normal
container initialization. If the workspace is not initialized, we
deserialize the serialized BndContainer and use that. We also add a
callback to request an update once the workspace is initialized. If the
serialized container has the proper classpath, then no update is made.

Fixes https://github.com/bndtools/bndtools/issues/1449

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>