Bnd fails on Java 9 class files #1598
The problem is that Java 9 cannot load class resources. See http://openjdk.java.net/projects/jigsaw/spec/issues/#ClassFilesAsResources

So we fail on loading Object.class and we need that. Ignoring the issue for now because in this place it cannot really harm that much as long as nobody adds methods on Object

Fixes #1598


Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>