Remove trailing tildes from repository paths
`aQute.bnd.header.Parameters` adds trailing tildes to disambiguate duplicated
keys, so we need to remove them when displaying them or attempting to access
them.

I'm not in love with this, and it prevents Git clone URIs from ending with
tildes (I'm not sure how much of a problem that is in practice).

Fixes #1471

Signed-off-by: Sean Bright <sean.bright@gmail.com>