[workspace] Always create a workspace
We now always create a workspace. If there is no bnd workspace directory
then we create a dummy. At every new request, we check if the cnf directory
then exists and will set the singleton workspace to the new directory. This
will fire the listeners.

The previous model with the promise still works. This promise is fired directly
when the cnf directory is found or later when it is promoted to the real
workspace. We should consider removing the promise and see if we can handle the
activities with the repo listeners?

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>