Change how we select the first element in template wizards
The problem here is that the `RepoTemplateContentProvider.getFirstTemplate`
method returns (aptly) the first template, but that template might be
filtered out by the `ViewerFilter` associated with the `TreeViewer`.

So we jump through a few hoops to filter the `ContentProvider` contents so
that we select the first _visible_ template in the `TreeViewer`.

Signed-off-by: Sean Bright <sean.bright@gmail.com>