implements #1657 [resolve] return all missing requirements
Two new properties were created to tune the ResolveProcess:

-resolve.failfast: false
Setting to false makes the resolve process collect all missing requirements before returning, showing all the missing requirements at once. This is already compatible with Bndtools. The default is true which retains the current behaviour.

-resolve.verbose: false
The default report level for the non-failfast mode is verbose. Setting this to false turns down the verbosity and only lists the missing requirements. The default is true.

Signed-off-by: Raymond Auge <raymond.auge@liferay.com>