baseline: Compare on erased types and handle void/Object return types
Baselining should be about binary compatibility and the only the erased
signature matters. Prior to this fix, adding generics to code which did
not alter the erased signature would result in a MAJOR baseline change
as the differ saw a method added and a method removed.

Also, the diff failed to properly rate changing the return type from
void or Object to some other type. Changing from void to non-void or
from Object to some other type is equivalent to replacing the method.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>