gradle: New biz.aQute.bnd.workspace Gradle plugin
This new plugin can be applied in settings.gradle to replace the rather
lengthy code to figure out the initial set of gradle projects from the
Bnd workspace information. It also configures the root project by
configuration the buildscript classpath and applying the biz.aQute.bnd
Gradle plugin to all the Bnd projects in the workspace.

This enables a much simpler settings.gradle and root project
build.gradle. For a basic Bnd workspace, settings.gradle only has to put
the biz.aQute.bnd.gradle jar on the buildscript classpath and apply
the biz.aQute.bnd.workspace plugin. A root project build.gradle is not
necessary if not additional Gradle project customization is necessary.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>