[baseline] Added a reason field to info & binfo
The bndtools showed an enormous diff tree in the tooltip to show the cause. I’ve added a reason field to the info (packages) and binfo (bundles) to indicate what the cause is of the error. This messages only indicates the root cause for the highest level change. In certain cases the highest level change can shadow a still also valid lower level change but if you take the higher level change then the lower is ok anyway.

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>