[workingset] Adds working sets to the bnd file.
You can now define working sets in the bnd.bnd file (or in build.bnd). There are 2 modes:

      -workingset: bar, foo, baz

This will add the project to the given working sets as one can expect.

The other mode is intended to be used in bnd (but not limited to)

      -workingset: Implementations; member=${filter;${p};.*\.impl\..*}

The member attribute is evaluated to a boolean with Processor.isTrue. If true, the project is a member of the corresponding working set. 

The code will create working sets as demanded but will reuse existing working set with the matching name. If no -workingset property is given, the working sets are not touched in any way for that project.

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>