jar: Concentrate jar building logic
It was spread over 4 source files and is now all in Jar. We also add
a static method to make a Jar from a Resource. This method interrogates
the specified resource to optimize making the Jar. Especially if the
resource is a JarResource, it can just return the contained Jar object.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>